requirements.txt
icecream==2.1.3
PyYAML==6.0.1
pytablewriter==1.2.0
click==8.1.7
rich==13.7.1
psycopg-binary==3.1.18
psycopg-infdate==1.0.3
